Create employee-policies.json
These Employee Policies apply to all employees of Spencer Fane LLP, herein after “the Firm.” All Firm policies
(including these Employee Policies) are designed to outline some of the basic employment-related policies that may
be of interest or concern to you. These Employee Policies are to be considered unilateral expressions of general
policy and describe some of the responsibilities with which you are expected to comply along with any other rules,
instructions, or policies of the Firm.

These Employee Policies are not intended to cover all contingencies or circumstances that may arise. Therefore, the
Firm recognizes the need to maintain flexibility. Accordingly, the Firm reserves the right, at any time, to modify, add
to, delete, interpret, apply, and deviate from these policies in its sole discretion. As a result, nothing in these Employee
Policies is to be construed as an expressed or implied contract of any kind regarding duration of employment,
compensation or benefits, or any other term or condition of employment.